+++ app/app_upgrades.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/archway-network/archway/app/upgrades"
 	upgrade_0_6 "github.com/archway-network/archway/app/upgrades/06"
+	upgrade10_0_0 "github.com/archway-network/archway/app/upgrades/10_0_0"
 	upgrade1_0_0_rc_4 "github.com/archway-network/archway/app/upgrades/1_0_0_rc_4"
 	upgrade2_0_0 "github.com/archway-network/archway/app/upgrades/2_0_0"
 	upgrade3_0_0 "github.com/archway-network/archway/app/upgrades/3_0_0"
@@ -14,7 +15,8 @@ import (
 	upgrade4_0_2 "github.com/archway-network/archway/app/upgrades/4_0_2"
 	upgrade6_0_0 "github.com/archway-network/archway/app/upgrades/6_0_0"
 	upgrade7_0_0 "github.com/archway-network/archway/app/upgrades/7_0_0"
-	upgrade8_0_0 "github.com/archway-network/archway/app/upgrades/8_0_0"
+	upgrade9_0_0 "github.com/archway-network/archway/app/upgrades/9_0_0"
+	"github.com/archway-network/archway/app/upgrades/constantineupgrades"
 )
 
 // UPGRADES
@@ -28,7 +30,12 @@ var Upgrades = []upgrades.Upgrade{
 	upgrade4_0_2.Upgrade,      // v4.0.2
 	upgrade6_0_0.Upgrade,      // v6.0.0
 	upgrade7_0_0.Upgrade,      // v7.0.0
-	upgrade8_0_0.Upgrade,      // v8.0.0
+	// upgrade8_0_0.Upgrade,  // v8.0.0: was reserved for a consensus breaking wasmd upgrade
+	upgrade9_0_0.Upgrade,  // v9.0.0
+	upgrade10_0_0.Upgrade, // v10.0.0
+
+	// constantine only
+	constantineupgrades.WASMD_50_Amino_Patch,
 }
 
 func (app *ArchwayApp) RegisterUpgradeHandlers() {